add TemperatureFormatter class

+++ task9mvvm/src/main/java/by/itacademy/training/task9mvvm/model/entity/CurrentTemperature.kt
@@ -1,6 +1,6 @@
 package by.itacademy.training.task9mvvm.model.entity
 
 data class CurrentTemperature(
-    val celsiusTemperature: Double,
-    val fahrenheitTemperature: Double,
+    val celsiusTemperature: Int,
+    val fahrenheitTemperature: Int,
 )
