+++ task9mvvm/src/main/java/by/itacademy/training/task9mvvm/model/entity/HourTemperature.kt
@@ -1,7 +1,7 @@
 package by.itacademy.training.task9mvvm.model.entity
 
 data class HourTemperature(
-    val celsiusTemperature: Double,
-    val fahrenheitTemperature: Double,
+    val celsiusTemperature: Int,
+    val fahrenheitTemperature: Int,
     val time: String,
 )
